refactor: split pages/ into different subfolders
