+++ learnhub-ui/src/pages/home/Home.tsx
@@ -1,7 +1,7 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../hooks/useAuth";
-import { UserRole } from "../types/Account";
-//import StudentHome from "./StudentHome";
+import { useAuth } from "../../hooks/useAuth";
+import { UserRole } from "../../types/Account";
+import StudentHome from "./StudentHome";
 import TeacherHome from "./TeacherHome";
 import TeacherManagerHome from "./TeacherManagerHome";
 
@@ -9,7 +9,7 @@ export default function Home() {
     const { account } = useAuth();
     switch (account?.role) {
         case UserRole.STUDENT:
-            return <TeacherManagerHome />;
+            return <StudentHome />;
         case UserRole.TEACHER:
             return <TeacherHome />;
         case UserRole.TEACHER_MANAGER:
