+++ learnhub-ui/src/pages/home/teacher/TeacherHome.tsx
@@ -1,4 +1,4 @@
-import { HomeLayout } from "../layouts";
+import { HomeLayout } from "../../../layouts";
 
 export default function TeacherHome() {
     return (
