+++ learnhub-ui/src/pages/home/teachermanager/TeacherManagerHome.tsx
@@ -1,4 +1,4 @@
-import { HomeLayout } from "../layouts";
+import { HomeLayout } from "../../../layouts";
 
 export default function TeacherManagerHome() {
     return (
