+++ learnhub-ui/src/pages/index.ts
@@ -1,14 +1,14 @@
-export { default as Landing } from "./Landing";
-export { default as Login } from "./Login";
-export { default as ManagerLogin } from "./ManagerLogin";
-export { default as Register } from "./Register";
-export { default as NotFound } from "./NotFound";
-export { default as Unauthorized } from "./Unauthorized";
-export { default as ForgotPassword } from "./ForgotPassword";
-export { default as ResetPassword } from "./ResetPassword";
-export { default as UserProfile } from "./UserProfile";
-export { default as TeacherDetail } from "./TeacherDetail";
-export { default as CourseList } from "./CourseList";
-export { default as ContactUs } from "./ContactUs";
-export { default as Home } from "./Home";
-export { default as Mailbox } from "./Mailbox";
+export { default as Landing } from "./public/Landing";
+export { default as Login } from "./auth/Login";
+export { default as ManagerLogin } from "./auth/ManagerLogin";
+export { default as Register } from "./auth/Register";
+export { default as NotFound } from "./error/NotFound";
+export { default as Unauthorized } from "./error/Unauthorized";
+export { default as ForgotPassword } from "./auth/ForgotPassword";
+export { default as ResetPassword } from "./auth/ResetPassword";
+export { default as UserProfile } from "./home/UserProfile";
+export { default as TeacherDetails } from "./public/TeacherDetails";
+export { default as CourseList } from "./public/CourseList";
+export { default as ContactUs } from "./public/ContactUs";
+export { default as Home } from "./home/Home";
+export { default as Mailbox } from "./home/teachermanager/Mailbox";
